Target netstandard20 instead
Discord.Net is linked against an old version of
System.Threading.Tasks.Extensions that is not targeted to net461. The
Nuget versioning allows pulling in a newer version of
System.Threading.Tasks.Extensions that is targeted against net461, but
any such version of Tasks.Extensions has an internal version of 4.2.0.1,
which matches 4.2.0.0 in Mono, but not on Windows.